Added a test to check encoding against DER specification
This test checks the encoding of several ASN1 primitives to the encoding in DER. All existing types that I could find a test case for (Integer, BitString and OctetString) pass the test. I'm pretty sure ASN1Length is also DER-compliant, so that would mean Sequence as well. Which means basically everything is.

I also created a test in Dartcoin to test for signature encoding to DER and it passed using asn1lib.

Involves #3